+++ NEWS.md
@@ -6,6 +6,7 @@ the attendant instead of `httr::progress`.
 `incProgressWairess` analogous to the same functions in shiny.
 - Added `withProgressAttendant`,
 `incProgressAttendant`, and `setProgressAttendant`.
+- Fix major issue with `Hostess`, it was basically broken.
 
 # waiter 0.2.3