+++ NEWS.md
@@ -6,7 +6,8 @@ the attendant instead of `httr::progress`.
 `incProgressWairess` analogous to the same functions in shiny.
 - Added `withProgressAttendant`,
 `incProgressAttendant`, and `setProgressAttendant`.
-- Fix major issue with `Hostess`, it was basically broken.
+- Fix major issue with `Hostess`, it was basically broken
+[#108](https://github.com/JohnCoene/waiter/issues/108).
 
 # waiter 0.2.3