+++ NEWS.md
@@ -1,4 +1,4 @@
-# waiter 0.2.4.9000
+# waiter 0.2.4
 
 - Added `httr_progress` function to use  the waitress or
 the attendant instead of `httr::progress`.